update tests for snapshot verification
